♻️ Refactor : 화면 단위 컴포넌트로 파일 구조 수정
