+++ src/features/onboarding/components/PaginationView.tsx
